+++ CHANGELOG.md
@@ -2,6 +2,14 @@ GAMSPy CHANGELOG
 ================
 
 -------------------------------------------------------------------------------
+GAMSPy 0.11.8
+-------------------------------------------------------------------------------
+- General
+  - Allow assigning VariableType enum or EquationType enum as an attribute after object creation for Equation and Variable.
+- Testing
+  - Add tests for assigning type to Variable and Equation after creation.
+
+-------------------------------------------------------------------------------
 GAMSPy 0.11.7
 -------------------------------------------------------------------------------
 - General